Make the Linux day compile: eight errors, six of them one lesson
Nine commits written without a compiler produced eight errors, and six were
two mistakes repeated:

Implicit returns. Passage.afterword and Stray.Stage.x(in:) both put a
statement before their switch, which makes the switch a statement rather
than an expression — Swift stops implying the returns and says nothing until
a compiler sees it. A brace-tracking scan found every other instance in the
app; the other three hits were assignments and are fine.

Exhaustiveness against enums that grew today. ChronicleEvent.Kind reached
fifteen cases and YearRing's rim handled nine; CatalogItem reached six and
UnlockSheet's portrait drew three; Dream reached twenty-one and
daysSinceMeeting matched thirteen. The rim and the portrait needed real
decisions rather than fall-through, so residents, first nights and the
hundred-hour wood now mark the rim, and accessories, dens and film stocks
have portraits — a film stock shown by grading a real scene, since what it
is is what it does to a picture.

And three of their own: Heard.Spec had a defaulted field declared mid-struct
so every row supplying it had to thread it through the middle (moved last),
BuddySprite was called with outfit before sleeping, a local ShelfOfHours
array shadowed the engine's Shelf store, and the keepsake gate tested the
bond level just crossed — usually nil — instead of the level reached.

+++ Pawmodoro/Views/YearRingView.swift
@@ -135,8 +135,8 @@ struct YearRingView: View {
                 )
                 let radius = outer + 4
                 let point = CGPoint(
-                    x: centre.x + cos(angle.radians) * radius,
-                    y: centre.y + sin(angle.radians) * radius
+                    x: centre.x + CGFloat(cos(angle.radians)) * radius,
+                    y: centre.y + CGFloat(sin(angle.radians)) * radius
                 )
                 context.fill(
                     Path(ellipseIn: CGRect(x: point.x - 2, y: point.y - 2,
